requirements.txt
# ==============================================================================
# MediAssist Pro — Python dependencies
# Pinned for reproducible Docker builds.
# ==============================================================================

# ── Core API framework ─────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
python-multipart==0.0.9

# ── LLM client ────────────────────────────────────────────────────────────────
groq==0.9.0

# ── LangChain (orchestration + RAG) ───────────────────────────────────────────
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.3
langchain-chroma==0.1.1

# ── Vector database ────────────────────────────────────────────────────────────
chromadb==0.5.0

# ── Embeddings ─────────────────────────────────────────────────────────────────
sentence-transformers==3.0.0

# ── Fine-tuning (optional — comment out if running inference-only) ─────────────
transformers==4.41.0
datasets==2.19.1
peft==0.11.1
accelerate==0.30.1
trl==0.8.6
# bitsandbytes==0.43.1   # uncomment for QLoRA (CUDA only)

# ── Data processing ────────────────────────────────────────────────────────────
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2

# ── Utilities ─────────────────────────────────────────────────────────────────
httpx==0.27.0
tenacity==8.3.0
loguru==0.7.2

# ── Testing ────────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
